Add anchors for relevant manpage HTML tags
This patch adds a new `add_achors.js` script, which can be executed in
the browser console to add IDs to all major tags in Alacritty's manpage
HTML documents. This is a very manual process, but it should be fine
considering there's only 4 files which need this treatment.

Since it came up in an issue before, this patch also goes ahead and
pulls the other two manpages from Alacritty's repo. Since the conversion
is mostly automated this is easy to provide, so it can be useful as a
quick reference.

Closes #29.